Patch: Stable FPS for logic function of game-loop
This has been done by applying a delta value by which
other delta properties are adjusted. I.e. the moveVelocity
of the player entity at a given loop iteration will
be increased if the frame-rate is low or decreased if the
frame-rate is high. This provides a steady logic update
that is not affected by the fps of the game. In other words,
the game's performance can drop without affecting the
outcome of its logic.

	modified:   util.js